v2.4.5.3 -> v2.4.5.4
  - Chris Mason: ReiserFS pre-allocation locking bugfix
  - David Miller: fix bitops users (requires "long" alignment)
  - Andrey Savochkin: file locking failure case SMP lock fix
  - Urban Widmark: smbfs update (avoid unnecessary flushing, make NetApp
  work)
  - Andrew Grover: ACPI update
  - Jeff Garzik: network driver updates
  - Maciej Rozycki: IO-APIC level trigger problem workaround
  - Rusty Russell: ipt_unclean fix
  - Richard Gooch: devfs update

+++ drivers/ide/ide-floppy.c
@@ -237,7 +237,7 @@ typedef struct {
 	idefloppy_flexible_disk_page_t flexible_disk_page;	/* Copy of the flexible disk page */
 	int wp;							/* Write protect */
 
-	unsigned int flags;			/* Status/Action flags */
+	unsigned long flags;			/* Status/Action flags */
 } idefloppy_floppy_t;
 
 /*
